Added DB migration
